Use group router
Group router deals with the receptionist for us, but it looks like the
node should join the cluster prior to sending messages through the
group router, or else the messages sent get dropped. The ugly
Thread.sleep is there for now to give time for the node to join the
cluster, but learning Cluster API should provide elegant ways of doing
this.