Set up testing, make first test pass
RareBooksProtocol.scala code is largely copy and paste, but the testing
setup is revised for ScalaTest 3.1.x, and simplified to begin with.
In the book's code, the common project has BaseSpec, BaseAkkaSpec
defined and extended from, but this may not be necessary these days, as
Akka has a ScalaTestWithActorTestKit provided.
This is not used here, because we aren't testing Actors yet.