Refactor message protocol
RareBooks needs to be able to handle not only its internal messages,
that are generated automatically by timers, but also the
RareBooksProtocol itself, which is a protocol shared amongst the
collaborating actors in this module.

As per the style guide, there are different options:
https://doc.akka.io/docs/akka/current/typed/style-guide.html#where-to-define-messages
Because we are using a shared protocol object, the approach taken here
was to modify the type hierachy so that the RareBooks internal messages
could become subclasses of it.
The RareBooks apply() method is made to return an ActorRef that can be
sent only the public RareBooksProtocol messages, but internally the
setup method actually creates the Behavior such that it can receive
messages of the entire type hierachy, including the private subtypes
that are used for RareBooks' internal messaging to itself.

The RareBookSpec tests hook directly into the setup method by entering a
"backdoor" version of apply, which does not call "narrow", enabling
tests to include sending of RareBooks' private messages to a test
instance.