Add replyTo ActorRefs to RareBooksProtocol
Besides the changes to the protocol messages, changes are needed to
the testing of the RareBooksProtocol messages, as well as the test suites
that use the protocol messages.

In Classic Akka Actors, the sender of messages could be obtained from
the context by called sender(); in Akka Typed Actors this is not
possible and it is required to explicitly add a replyTo ActorRef to
those messages expecting responses, so that the type of the ActorRef can
be specified to the compiler.