Add timestamp to state
Just to be consistent with the data structure from the book. The book
appears to not use the timestamp anywhere so it's kind of moot.
Also not sure the point of copying the timestamp from the BookFound
message into the state either, because I imagine that a Customer could
receive responses from different Librarians out-of-sequence, which would
update the BookFound count, but give an older timestamp in the state.