Use supervision to keep the show going
I don't really like this, but the scenario in this section of the
original book models a Librarian getting frustrated by throwing an
Exception.

In Akka Classic supervision, the supervision can get its
hands on the details of the Exception thrown and take action based on
that.

However this is not possible in Akka Typed supervison; only the type of
exception thrown is available to the superviser, and the supervisor can
not take actions such as assuming the responsibility of the faulty
actor.

But to play along with the book examples further, this is what I've
done.